+++ CHANGELOG.md
@@ -5,12 +5,14 @@
 ### Modules
 * `icon_theme`: prints the current icon theme for GTK
 * `cursor_theme`: prints the current cursor theme for GTK
+* `battery`: prints the current battery percentage and status
 
 ### Config syntax
 * any invalid escape sequence like "\X" will now just be reduced to "X"
 * it is now possible to include "#" and ";" in your config without them counting as a comment, as long as they are enclosed between two `"` (which should always be true anyway)
 * `col_block_len` was removed, in favor of `col_block_str`, which contains the entire string that is printed for each color
 * `col_background` was added, which controls whether the background or the background of `col_block_str` is colored (vs. the foreground)
+* `bat_status` toggles whether the battery status is shown
 
 ### Command line arguments